hw5 changes to work
