+++ src/main/java/com/example/amazoncom/utils/HelperClass.java
@@ -1,7 +1,10 @@
+package com.example.amazoncom.utils;
+
 import io.github.bonigarcia.wdm.WebDriverManager;
 import java.time.Duration;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
 import org.openqa.selenium.support.ui.WebDriverWait;
 
 public class HelperClass {
@@ -12,17 +15,13 @@ public class HelperClass {
   private static WebDriverWait wait;
 
   private HelperClass() {
-
     WebDriverManager.chromedriver().setup();
-    driver = new ChromeDriver();
+    ChromeOptions chromeOptions = new ChromeOptions();
+    chromeOptions.addArguments("--remote-allow-origins=*");
+    driver = new ChromeDriver(chromeOptions);
     wait = new WebDriverWait(driver, Duration.ofSeconds(TIMEOUT));
     driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(TIMEOUT));
     driver.manage().window().maximize();
-
-  }
-
-  public static void openPage(String url) {
-    driver.get(url);
   }
 
   public static WebDriver getDriver() {
@@ -31,21 +30,16 @@ public class HelperClass {
   }
 
   public static void setUpDriver() {
-
     if (helperClass == null) {
-
       helperClass = new HelperClass();
     }
   }
 
   public static void tearDown() {
-
     if (driver != null) {
       driver.close();
       driver.quit();
     }
-
     helperClass = null;
-
   }
 }
