+++ src/test/java/com/example/amazoncom/testRunners/CucumberRunnerTests.java
@@ -1,4 +1,4 @@
-package testRunners;
+package com.example.amazoncom.testRunners;
 
 import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
